Prepend `openapi!` in any case
If you run tests with Minitest at themoment, but without the environment variable `OPENAPI`, you are greeted with an error message.

`undefined method `openapi!' for Api::V0::SomeControllerTest:Class (NoMethodError)`

This is because the current code adds the `openapi!` method only if `OPENAPI` environment variable is set.

This commit modifies the loading logic, always requiring the `minitest` hooks class to be loaded if `Minitest` constant is defined, to avoid issues with RSpec. Then, the `openapi?` and `openapi!` methods are always prepended to `Minitest::Test`, while the modified `run` method and the `after_run` hook are only activated when `OPENAPI` environment variable is set.